+++ packages/expo-file-system/src/next/FileSystem.ts
@@ -1,5 +1,3 @@
-import { ReadableStream, WritableStream } from 'web-streams-polyfill';
-
 import ExpoFileSystem from './ExpoFileSystem';
 import { PathUtilities } from './pathUtilities';
 import { FileSystemReadableStreamSource, FileSystemWritableSink } from './streams';