+++ packages/@expo/metro-config/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Shim web streams when bundling for the server.
+
 ## 0.20.7 — 2025-04-25
 
 _This version does not introduce any user-facing changes._